Extract User into a separate module
