+++ test/EvaluationLogTests.ts
@@ -5,8 +5,9 @@ import * as util from "util";
 import "mocha";
 import { User } from "../src";
 import { LogLevel, LoggerWrapper } from "../src/ConfigCatLogger";
-import { SettingValue, WellKnownUserObjectAttribute } from "../src/ProjectConfig";
+import { SettingValue } from "../src/ProjectConfig";
 import { RolloutEvaluator, evaluate } from "../src/RolloutEvaluator";
+import { WellKnownUserObjectAttribute } from "../src/User";
 import { errorToString } from "../src/Utils";
 import { CdnConfigLocation, ConfigLocation, LocalFileConfigLocation } from "./helpers/ConfigLocation";
 import { FakeLogger } from "./helpers/fakes";
